Code refactor and files name change
