+++ src/components/atoms/ProfileTitleBar.tsx
@@ -5,7 +5,7 @@ import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
 import {useDispatch} from 'react-redux';
 import colors from '../../assets/theme/colors';
 import {SETTINGS} from '../../models/constants/routeNames';
-import {logOutUser} from '../../redux/userThunk/UserAction';
+import {logOutUser} from '../../redux/user/UserAction';
 
 const ProfileTitleBar = () => {
   const {navigate} = useNavigation();
